+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Navigation arrows would not work when the `slider` was placed inside an `<a>` tag. This is the case if you were to place a `slider` inside of a `product-summary`.
 
 ## [1.44.0] - 2020-09-22
 ### Added