+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- List name to GTM `productClick` event.
+- Passing `listName` prop to Product Summary.
 
 ## [1.44.2] - 2020-12-01
 ### Fixed