+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - List name to GTM `productClick` event.
-- Passing `listName` prop to Product Summary.
+- Passing `listName` and `position` prop to Product Summary.
 
 ### Fixed
 - `trackingId` is now being used on `RelatedProducts` shelf.