+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `clusterHighlights` to `ProductRecommendations` query.
+
 ## [1.45.0] - 2021-06-09
 ### Added
 - List name to GTM `productClick` event.