+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - `hideOutOfStockItems` prop to `RelatedProducts` shelf
+
+## [1.46.0] - 2021-09-06
+
+### Added
+- `clusterHighlights` to `ProductRecommendations` query.
+
 ## [1.45.0] - 2021-06-09
 ### Added
 - List name to GTM `productClick` event.